Text.Blaze.Renderer.XmlHtml

Description

Renderer that supports rendering to xmlhtml forests. This is a port of the Hexpat renderer.

Warning: because this renderer doesn't directly create the output, but rather an XML tree representation, it is impossible to render pre-escaped text. This means that preEscapedString will produce the same output as string. This also applies to the functions preEscapedText, preEscapedTextValue...

Documentation

renderHtml :: Html -> Document

Render HTML to an xmlhtml Document

renderHtmlNodes :: Html -> [Node]

Render HTML to a list of xmlhtml nodes